+++ tests/test_40_xarray_encoders.py
@@ -10,6 +10,7 @@ try:
     import xarray as xr
 except ImportError:
     pytest.importorskip("xarray")
+    pytest.importorskip("dask")
 
 T = TypeVar("T")
 
@@ -105,57 +106,3 @@ def test_xr_cacheable(ds: xr.Dataset, xarray_cache_type: str, extension: str) ->
             xr.testing.assert_equal(res, ds)
         else:
             xr.testing.assert_identical(res, ds)
-
-
-def test_dictify_io_object(tmpdir: str) -> None:
-    tmpfile = os.path.join(tmpdir, "dummy.txt")
-    with open(tmpfile, "w") as f:
-        f.write("dummy")
-
-    local_path = os.path.join(
-        config.SETTINGS["cache_store"].directory,
-        "f6e6e2cc3b79d2ff7163fe28e6324870bfe8cf16a912dfc2ebceee7a.txt",
-    )
-    expected = {
-        "type": "text/plain",
-        "href": "./f6e6e2cc3b79d2ff7163fe28e6324870bfe8cf16a912dfc2ebceee7a.txt",
-        "file:checksum": "f6e6e2cc3b79d2ff7163fe28e6324870bfe8cf16a912dfc2ebceee7a",
-        "file:size": 5,
-        "file:local_path": local_path,
-        "io:open_kwargs": {"encoding": "UTF-8", "errors": "strict", "mode": "r"},
-    }
-    res = extra_encoders.dictify_io_object(open(tmpfile))
-    assert res == expected
-    assert os.path.exists(local_path)
-
-
-def test_copy_file_to_cache_directory(tmpdir: str) -> None:
-    tmpfile = os.path.join(tmpdir, "dummy.txt")
-    cached_file = os.path.join(
-        tmpdir, "6b4e03423667dbb73b6e15454f0eb1abd4597f9a1b078e3f5b5a6bc7.txt"
-    )
-
-    with open(tmpfile, "w") as f:
-        f.write("dummy")
-    cfunc = cache.cacheable(func)
-
-    res = cfunc(open(tmpfile))
-    assert res.read() == "dummy"
-    with open(cached_file, "r") as f:
-        assert f.read() == "dummy"
-    assert config.SETTINGS["cache_store"].stats() == (0, 1)
-
-    # skip copying a file already in cache directory
-    mtime = os.path.getmtime(cached_file)
-    res = cfunc(open(tmpfile))
-    assert res.read() == "dummy"
-    assert mtime == os.path.getmtime(cached_file)
-    assert config.SETTINGS["cache_store"].stats() == (1, 1)
-
-    # do not crash if cached file is removed
-    os.remove(cached_file)
-    with pytest.warns(UserWarning):
-        res = cfunc(open(tmpfile))
-    assert res.read() == "dummy"
-    assert os.path.exists(cached_file)
-    assert config.SETTINGS["cache_store"].stats() == (2, 1)
